pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.19.5",
    "pandas>=2.2.3",
    "streamlit-folium>=0.24.0",
    "streamlit>=1.44.0",
    "numpy>=2.2.4",
    "plotly>=6.0.1",
    "pillow>=11.1.0",
    "streamlit-extras>=0.6.0",
    "components>=0.0.1a0",
    "reportlab>=4.3.1",
    "xlsxwriter>=3.2.2",
]
